dn_llm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dn_llm_czlabinger"
version = "0.0.1"
authors = [
  { name="czlabinger", email="czlabinger@student.tgm.ac.at" },
]
description = "Package to interact with LLMs"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
	"llama-cpp-python"
]
